temporary commit, move popcount, reduce and active sse splits
